+++ src/lib/components/table/codes/CodesTableHeader.tsx
@@ -1,0 +1,18 @@
+import type { GridProps } from "@chakra-ui/react";
+import { Grid } from "@chakra-ui/react";
+
+import { TableHeader } from "lib/components/table";
+
+export const CodesTableHeader = ({
+  templateColumns,
+}: {
+  templateColumns: GridProps["templateColumns"];
+}) => (
+  <Grid templateColumns={templateColumns} minW="min-content">
+    <TableHeader borderTopStyle="none">Code ID</TableHeader>
+    <TableHeader>Code Name</TableHeader>
+    <TableHeader>Contracts</TableHeader>
+    <TableHeader>Uploader</TableHeader>
+    <TableHeader>Permission</TableHeader>
+  </Grid>
+);