+++ src/lib/components/table/proposals/index.tsx
@@ -1,0 +1,2 @@
+export * from "./ProposalsTableHeader";
+export * from "./ProposalsTableRow";