+++ src/lib/pages/account-details/components/asset/index.tsx
@@ -96,7 +96,7 @@ export const AssetsSection = ({
       />
       {supportedAssets &&
         onViewMore &&
-        supportedAssets.length >= MaxAssetsShow && (
+        supportedAssets.length > MaxAssetsShow && (
           <ViewMore onClick={onViewMore} />
         )}
     </Flex>